+++ src/infrastructure/utils/controllers.ts
@@ -1,39 +1,39 @@
-import { AdminController } from "../adapters/controllers/adminController"
-import { ChatController } from "../adapters/controllers/chatController"
-import { CouponController } from "../adapters/controllers/couponController"
-import { MovieController } from "../adapters/controllers/movieController"
-import { ScreenController } from "../adapters/controllers/screenController"
-import { ScreenSeatController } from "../adapters/controllers/screenSeatController"
-import { ShowController } from "../adapters/controllers/showController"
-import { ShowSeatController } from "../adapters/controllers/showSeatsController"
-import { TheaterController } from "../adapters/controllers/theaterController"
-import { TicketController } from "../adapters/controllers/ticketController"
-import { UserController } from "../adapters/controllers/userController"
-import { AdminRepository } from "../infrastructure/repositories/adminRepository"
-import { ChatRepository } from "../infrastructure/repositories/chatRepository"
-import { CouponRepository } from "../infrastructure/repositories/couponRepository"
-import { MovieRepository } from "../infrastructure/repositories/movieRepository"
-import { ScreenRepository } from "../infrastructure/repositories/screenRepository"
-import { ScreenSeatRepository } from "../infrastructure/repositories/screenSeatRepository"
-import { ShowRepository } from "../infrastructure/repositories/showRepository"
-import { ShowSeatsRepository } from "../infrastructure/repositories/showSeatRepository"
-import { TempTheaterRepository } from "../infrastructure/repositories/tempTheaterRepository"
-import { TempTicketRepository } from "../infrastructure/repositories/tempTicketRepository"
-import { TempUserRepository } from "../infrastructure/repositories/tempUserRepository"
-import { TheaterRepository } from "../infrastructure/repositories/theaterRepository"
-import { TicketRepository } from "../infrastructure/repositories/ticketRepository"
-import { UserRepository } from "../infrastructure/repositories/userRepository"
-import { AdminUseCase } from "../useCases/adminUseCase"
-import { ChatUseCase } from "../useCases/chatUseCase"
-import { CouponUseCase } from "../useCases/couponUseCase"
-import { MovieUseCase } from "../useCases/movieUseCase"
-import { ScreenSeatUseCase } from "../useCases/screenSeatUseCase"
-import { ScreenUseCase } from "../useCases/screenUseCase"
-import { ShowSeatsUseCase } from "../useCases/showSeatUseCase"
-import { ShowUseCase } from "../useCases/showUseCase"
-import { TheaterUseCase } from "../useCases/theaterUseCase"
-import { TicketUseCase } from "../useCases/ticketUseCase"
-import { UserUseCase } from "../useCases/userUseCase"
+import { AdminController } from "../../adapters/controllers/adminController"
+import { ChatController } from "../../adapters/controllers/chatController"
+import { CouponController } from "../../adapters/controllers/couponController"
+import { MovieController } from "../../adapters/controllers/movieController"
+import { ScreenController } from "../../adapters/controllers/screenController"
+import { ScreenSeatController } from "../../adapters/controllers/screenSeatController"
+import { ShowController } from "../../adapters/controllers/showController"
+import { ShowSeatController } from "../../adapters/controllers/showSeatsController"
+import { TheaterController } from "../../adapters/controllers/theaterController"
+import { TicketController } from "../../adapters/controllers/ticketController"
+import { UserController } from "../../adapters/controllers/userController"
+import { AdminRepository } from "../repositories/adminRepository"
+import { ChatRepository } from "../repositories/chatRepository"
+import { CouponRepository } from "../repositories/couponRepository"
+import { MovieRepository } from "../repositories/movieRepository"
+import { ScreenRepository } from "../repositories/screenRepository"
+import { ScreenSeatRepository } from "../repositories/screenSeatRepository"
+import { ShowRepository } from "../repositories/showRepository"
+import { ShowSeatsRepository } from "../repositories/showSeatRepository"
+import { TempTheaterRepository } from "../repositories/tempTheaterRepository"
+import { TempTicketRepository } from "../repositories/tempTicketRepository"
+import { TempUserRepository } from "../repositories/tempUserRepository"
+import { TheaterRepository } from "../repositories/theaterRepository"
+import { TicketRepository } from "../repositories/ticketRepository"
+import { UserRepository } from "../repositories/userRepository"
+import { AdminUseCase } from "../../useCases/adminUseCase"
+import { ChatUseCase } from "../../useCases/chatUseCase"
+import { CouponUseCase } from "../../useCases/couponUseCase"
+import { MovieUseCase } from "../../useCases/movieUseCase"
+import { ScreenSeatUseCase } from "../../useCases/screenSeatUseCase"
+import { ScreenUseCase } from "../../useCases/screenUseCase"
+import { ShowSeatsUseCase } from "../../useCases/showSeatUseCase"
+import { ShowUseCase } from "../../useCases/showUseCase"
+import { TheaterUseCase } from "../../useCases/theaterUseCase"
+import { TicketUseCase } from "../../useCases/ticketUseCase"
+import { UserUseCase } from "../../useCases/userUseCase"
 import { Encrypt } from "./bcryptPassword"
 import { JWTToken } from "./jwtToken"
 import { MailSender } from "./nodemailer"
